test: Reduce test noise level
This makes test more readable by reducing log noise.
