+++ pkg/judge/judge_test.go
@@ -1,0 +1,9 @@
+package judge
+
+import (
+	"github.com/doitintl/kube-no-trouble/test"
+)
+
+func init() {
+	test.Setup()
+}
